+++ mek_data_class_generator/lib/src/configs.dart
@@ -1,31 +1,30 @@
+import 'package:json_annotation/json_annotation.dart';
+
+part 'configs.g.dart';
+
+enum StringifyType { params, fields }
+
+@JsonSerializable()
 class Config {
   final int pageWidth;
   final bool comparable;
   final bool stringify;
+  final StringifyType stringifyType;
   final bool copyable;
   final bool changeable;
   final bool changesVisible;
   final bool createFieldsClass;
 
   const Config({
-    required this.pageWidth,
-    required this.comparable,
-    required this.copyable,
-    required this.stringify,
-    required this.changeable,
-    required this.changesVisible,
-    required this.createFieldsClass,
+    this.pageWidth = 80,
+    this.comparable = true,
+    this.stringify = true,
+    this.stringifyType = StringifyType.params,
+    this.copyable = false,
+    this.changeable = false,
+    this.changesVisible = false,
+    this.createFieldsClass = false,
   });
 
-  factory Config.from(Map<String, dynamic> map) {
-    return Config(
-      pageWidth: map['pageWidth'] ?? 80,
-      comparable: map['comparable'] ?? true,
-      stringify: map['stringify'] ?? true,
-      copyable: map['copyable'] ?? false,
-      changeable: map['changeable'] ?? false,
-      changesVisible: map['changesVisible'] ?? false,
-      createFieldsClass: map['createFieldsClass'] ?? false,
-    );
-  }
+  factory Config.fromJson(Map<String, dynamic> map) => _$ConfigFromJson(map);
 }