+++ mek_data_class_generator/lib/src/specs.dart
@@ -25,6 +25,7 @@ class ClassSpec {
   final List<String> types;
   final bool comparable;
   final bool stringify;
+  final StringifyType stringifyType;
   final bool copyable;
   final bool changeable;
   final bool changesVisible;
@@ -36,6 +37,7 @@ class ClassSpec {
     required this.comparable,
     required this.copyable,
     required this.stringify,
+    required this.stringifyType,
     required this.changeable,
     required this.changesVisible,
     required this.createFieldsClass,
@@ -71,6 +73,7 @@ class ClassSpec {
       types: element.typeParameters.map((e) => e.displayName).toList(),
       comparable: annotation.peek('comparable')?.boolValue ?? config.comparable,
       stringify: annotation.peek('stringify')?.boolValue ?? config.stringify,
+      stringifyType: config.stringifyType,
       copyable: annotation.peek('copyable')?.boolValue ?? config.copyable,
       changeable: annotation.peek('changeable')?.boolValue ?? config.changeable,
       changesVisible: annotation.peek('changesVisible')?.boolValue ?? config.changesVisible,
@@ -89,13 +92,15 @@ class FieldSpec {
   final bool updatable;
   final String? stringifier;
 
-  const FieldSpec({
+  late final bool isParam = _isParam(element.enclosingElement3 as ClassElement, element);
+
+  FieldSpec({
     required this.element,
     required this.name,
     required this.comparable,
     required this.stringify,
     required this.updatable,
-    this.stringifier,
+    required this.stringifier,
   });
 
   factory FieldSpec.from(ClassSpec classSpec, FieldElement element) {
@@ -125,5 +130,13 @@ class FieldSpec {
     final name = type.getDisplayString(withNullability: true);
     final shouldNotNullable = !nullable && name.endsWith('?');
     return shouldNotNullable ? name.substring(0, name.length - 1) : name;
+  }
+
+  static bool _isParam(ClassElement classElement, FieldElement element) {
+    if (element.isPrivate) return false;
+    if (element.hasInitializer) return false;
+
+    final constructorElement = classElement.unnamedConstructor ?? classElement.constructors.first;
+    return constructorElement.parameters.any((e) => e.name == element.name);
   }
 }